fix MSVCPP warning

+++ bochs/cpu/softfloat3e/include/softfloat_types.h
@@ -64,7 +64,7 @@ struct uint128_extra { struct uint128 v; uint64_t extra; };
 struct f16_t {
     uint16_t v;
 #ifdef __cplusplus
-    f16_t(uint64_t v16): v(v16) {}
+    f16_t(uint16_t v16): v(v16) {}
     operator uint16_t() const { return v; }
 #endif
 };
@@ -72,7 +72,7 @@ struct f16_t {
 struct f32_t {
     uint32_t v;
 #ifdef __cplusplus
-    f32_t(uint64_t v32): v(v32) {}
+    f32_t(uint32_t v32): v(v32) {}
     operator uint32_t() const { return v; }
 #endif
 };
